Allow to mix agentic declarative and programmatic API
